Moved tests from totalscore_test.go to results_test.go
